move Events and ContactState
