+++ packages/dht/src/dht/contact/ContactList.ts
@@ -1,6 +1,20 @@
 import { PeerID, PeerIDKey } from '../../helpers/PeerID'
 import EventEmitter from 'eventemitter3'
-import { Events, ContactState } from './Contact'
+
+export class ContactState<C> {
+    public contacted = false
+    public active = false
+    public contact: C
+
+    constructor(contact: C) {
+        this.contact = contact
+    }
+}
+
+export interface Events<C> {
+    contactRemoved: (removedContact: C, closestContacts: C[]) => void
+    newContact: (newContact: C, closestContacts: C[]) => void
+}
 
 export class ContactList<C extends { getPeerId: () => PeerID }> extends EventEmitter<Events<C>> {
